chore(config): move providers into module
